requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
alembic>=1.11.0
cryptography>=41.0.0  # Required for secure PyMySQL connections    
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.0
geopy>=2.3.0
folium>=0.14.0
osmnx>=1.3.0
networkx>=3.1
tqdm>=4.65.0